Issue #348: Package restructuring
* Moves Entry to cache utils
* Moves AbstractPubSubListenerActor to cluster utils
