+++ services/concierge/cache/src/main/java/org/eclipse/ditto/services/concierge/cache/update/PolicyCacheUpdateActor.java
@@ -17,9 +17,9 @@ import static java.util.Objects.requireNonNull;
 import java.util.Collections;
 
 import org.eclipse.ditto.model.enforcers.Enforcer;
-import org.eclipse.ditto.services.base.actors.AbstractPubSubListenerActor;
+import org.eclipse.ditto.services.utils.cluster.AbstractPubSubListenerActor;
 import org.eclipse.ditto.services.models.concierge.EntityId;
-import org.eclipse.ditto.services.models.concierge.cache.Entry;
+import org.eclipse.ditto.services.utils.cache.entry.Entry;
 import org.eclipse.ditto.services.utils.cache.Cache;
 import org.eclipse.ditto.signals.commands.policies.PolicyCommand;
 import org.eclipse.ditto.signals.events.policies.PolicyEvent;
