+++ services/concierge/enforcement/src/main/java/org/eclipse/ditto/services/concierge/enforcement/EnforcerRetriever.java
@@ -23,7 +23,7 @@ import java.util.function.Function;
 
 import org.eclipse.ditto.model.enforcers.Enforcer;
 import org.eclipse.ditto.services.models.concierge.EntityId;
-import org.eclipse.ditto.services.models.concierge.cache.Entry;
+import org.eclipse.ditto.services.utils.cache.entry.Entry;
 import org.eclipse.ditto.services.utils.cache.Cache;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
