+++ services/concierge/enforcement/src/main/java/org/eclipse/ditto/services/concierge/enforcement/PolicyOrAclEnforcerRetrieverFactory.java
@@ -19,7 +19,7 @@ import java.util.Map;
 
 import org.eclipse.ditto.model.enforcers.Enforcer;
 import org.eclipse.ditto.services.models.concierge.EntityId;
-import org.eclipse.ditto.services.models.concierge.cache.Entry;
+import org.eclipse.ditto.services.utils.cache.entry.Entry;
 import org.eclipse.ditto.services.utils.cache.Cache;
 import org.eclipse.ditto.signals.commands.policies.PolicyCommand;
 import org.eclipse.ditto.signals.commands.things.ThingCommand;
