+++ services/models/concierge/pom.xml
@@ -71,6 +71,7 @@
                     <parameter>
                         <excludes>
                             <exclude>org.eclipse.ditto.services.models.concierge.ConciergeMessagingConstants</exclude>
+                            <exclude>org.eclipse.ditto.services.models.concierge.cache.Entry</exclude>
                             <exclude>org.eclipse.ditto.services.models.concierge.batch.BatchStepCommandRegistry</exclude>
                             <exclude>org.eclipse.ditto.services.models.concierge.batch.BatchStepCommandResponseRegistry</exclude>
                             <exclude>org.eclipse.ditto.services.models.concierge.batch.BatchMappingStrategy</exclude>
