+++ services/utils/cache/src/main/java/org/eclipse/ditto/services/utils/cache/entry/NonexistentEntry.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.ditto.services.models.concierge.cache;
+package org.eclipse.ditto.services.utils.cache.entry;
 
 import java.util.NoSuchElementException;
