+++ services/utils/cache/src/main/java/org/eclipse/ditto/services/utils/cache/entry/package-info.java
@@ -11,4 +11,4 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 @org.eclipse.ditto.utils.jsr305.annotations.AllValuesAreNonnullByDefault
-package org.eclipse.ditto.services.models.concierge.cache;
+package org.eclipse.ditto.services.utils.cache.entry;
